centralised come of the controller code
