+++ server/createGroup/editGroupController.ts
@@ -12,6 +12,12 @@ import RescheduleSessionsView from './date/rescheduleSessionView'
 import RescheduleOtherSessionsForm from '../editSession/dateAndTime/rescheduleOtherSessionsForm'
 import CreateOrEditGroupWhenPresenter from './when/createOrEditGroupWhenPresenter'
 import CreateOrEditGroupWhenView from './when/createOrEditGroupWhenView'
+import CreateOrEditGroupSexPresenter from './sex/createOrEditGroupSexPresenter'
+import CreateOrEditGroupSexView from './sex/createOrEditGroupSexView'
+import CreateOrEditGroupCohortPresenter from './cohort/createOrEditGroupCohortPresenter'
+import CreateOrEditGroupCohortView from './cohort/createOrEditGroupCohortView'
+import CreateOrEditGroupCodePresenter from './code/createOrEditGroupCodePresenter'
+import CreateOrEditGroupCodeView from './code/createOrEditGroupCodeView'
 
 export default class EditGroupController extends BaseController {
   protected readonly primaryNavigationTab = PrimaryNavigationTab.Groups
@@ -164,6 +170,150 @@ export default class EditGroupController extends BaseController {
 
     const presenter = new RescheduleSessionsPresenter(groupId, updatedGroupDetails, false, formError)
     const view = new RescheduleSessionsView(presenter)
+    return this.renderPage(res, view)
+  }
+
+  async editGroupSex(req: Request, res: Response): Promise<void> {
+    const { groupId } = req.params
+    const { username } = req.user
+    let userInputData = null
+    let formError: FormValidationError | null = null
+
+    const groupDetails = await this.accreditedProgrammesManageAndDeliverService.getGroupDetailsById(username, groupId)
+    const editGroupSexData = await this.accreditedProgrammesManageAndDeliverService.getGroupSexDetails(
+      username,
+      groupId,
+    )
+    const selectedSex = editGroupSexData.radios.find(
+      (radio: { selected: boolean; value: string }) => radio.selected,
+    )?.value
+
+    const createGroupFormData = {
+      sex: selectedSex || groupDetails?.sex,
+    }
+
+    if (req.method === 'POST') {
+      const data = await new CreateOrEditGroupForm(req, '').createGroupSexData()
+
+      if (data.error) {
+        res.status(400)
+        formError = data.error
+        userInputData = req.body
+      } else {
+        await this.accreditedProgrammesManageAndDeliverService.updateGroup(username, groupId, {
+          sex: data.paramsForUpdate.sex,
+        })
+
+        req.session.createGroupFormData = {}
+        return res.redirect(`/group/${groupId}/group-details`)
+      }
+    }
+
+    const presenter = new CreateOrEditGroupSexPresenter(
+      formError,
+      createGroupFormData,
+      userInputData,
+      groupId,
+      groupDetails?.code,
+    )
+    const view = new CreateOrEditGroupSexView(presenter)
+    return this.renderPage(res, view)
+  }
+
+  async editGroupCohort(req: Request, res: Response): Promise<void> {
+    const { groupId } = req.params
+    const { username } = req.user
+    let userInputData = null
+    let formError: FormValidationError | null = null
+
+    const groupDetails = await this.accreditedProgrammesManageAndDeliverService.getGroupDetailsById(username, groupId)
+    const bffEditGroupCohortData = await this.accreditedProgrammesManageAndDeliverService.getBffEditGroupCohort(
+      username,
+      groupId,
+    )
+    const selectedCohort = bffEditGroupCohortData.radios.find(
+      (radio: { selected: boolean; value: string }) => radio.selected,
+    )?.value
+
+    const createGroupFormData = {
+      cohort: selectedCohort || groupDetails?.cohort,
+    }
+
+    if (req.method === 'POST') {
+      const data = await new CreateOrEditGroupForm(req, '').createGroupCohortData()
+
+      if (data.error) {
+        res.status(400)
+        formError = data.error
+        userInputData = req.body
+      } else {
+        await this.accreditedProgrammesManageAndDeliverService.updateGroup(username, groupId, {
+          cohort: data.paramsForUpdate.cohort,
+        })
+
+        req.session.createGroupFormData = {}
+        return res.redirect(`/group/${groupId}/group-details`)
+      }
+    }
+
+    const presenter = new CreateOrEditGroupCohortPresenter(
+      formError,
+      createGroupFormData,
+      userInputData,
+      groupId,
+      groupDetails?.code,
+    )
+    const view = new CreateOrEditGroupCohortView(presenter)
+    return this.renderPage(res, view)
+  }
+
+  async editGroupCode(req: Request, res: Response): Promise<void> {
+    const { groupId } = req.params
+    const { username } = req.user
+    let userInputData = null
+    let formError: FormValidationError | null = null
+
+    const groupDetails = await this.accreditedProgrammesManageAndDeliverService.getGroupDetailsById(username, groupId)
+    const createGroupFormData = {
+      groupCode: groupDetails?.code,
+    }
+
+    if (req.method === 'POST') {
+      let existingGroup = { code: '' }
+
+      if (req.body['create-group-code']) {
+        const matchingGroup = await this.accreditedProgrammesManageAndDeliverService.getGroupByCodeInRegion(
+          username,
+          req.body['create-group-code'],
+        )
+
+        existingGroup = matchingGroup?.id === groupId ? { code: '' } : { code: matchingGroup?.code || '' }
+      }
+
+      const data = await new CreateOrEditGroupForm(req, existingGroup.code).createGroupCodeData()
+
+      if (data.error) {
+        res.status(400)
+        formError = data.error
+        userInputData = req.body
+      } else {
+        await this.accreditedProgrammesManageAndDeliverService.updateGroup(username, groupId, {
+          groupCode: data.paramsForUpdate.groupCode,
+        })
+
+        req.session.createGroupFormData = {}
+        return res.redirect(`/group/${groupId}/group-details`)
+      }
+    }
+
+    const presenter = new CreateOrEditGroupCodePresenter(
+      formError,
+      createGroupFormData,
+      userInputData,
+      groupId,
+      createGroupFormData.groupCode,
+    )
+    const view = new CreateOrEditGroupCodeView(presenter)
     return this.renderPage(res, view)
   }
 }
