+++ server/routes/index.ts
@@ -3,9 +3,6 @@ import { type RequestHandler, Router } from 'express'
 import CaselistController from '../caselist/caselistController'
 import ChangeCohortController from '../cohort/changeCohortController'
 import CreateGroupController from '../createGroup/createGroupController'
-import EditGroupCodeController from '../createGroup/code/editGroupCodeController'
-import EditGroupCohortController from '../createGroup/cohort/editGroupCohortController'
-import EditGroupSexController from '../createGroup/sex/editGroupSexController'
 import EditSessionController from '../editSession/editSessionController'
 import GroupController from '../group/groupController'
 import GroupDetailsController from '../groupDetails/groupDetailsController'
@@ -52,9 +49,6 @@ export default function routes({ accreditedProgrammesManageAndDeliverService }: 
     accreditedProgrammesManageAndDeliverService,
   )
   const createGroupController = new CreateGroupController(accreditedProgrammesManageAndDeliverService)
-  const editGroupCodeController = new EditGroupCodeController(accreditedProgrammesManageAndDeliverService)
-  const editGroupCohortController = new EditGroupCohortController(accreditedProgrammesManageAndDeliverService)
-  const editGroupSexController = new EditGroupSexController(accreditedProgrammesManageAndDeliverService)
   const groupController = new GroupController(accreditedProgrammesManageAndDeliverService)
   const groupDetailsController = new GroupDetailsController(accreditedProgrammesManageAndDeliverService)
   const sessionScheduleController = new SessionScheduleController(accreditedProgrammesManageAndDeliverService)
@@ -241,16 +235,16 @@ export default function routes({ accreditedProgrammesManageAndDeliverService }: 
     await createGroupController.showCreateGroupCode(req, res)
   })
   getOrPost('/group/:groupId/edit-a-group/edit-group-code', async (req, res) => {
-    await editGroupCodeController.showEditGroupCode(req, res)
+    await editGroupController.editGroupCode(req, res)
   })
   getOrPost('/group/create-a-group/group-cohort', async (req, res) => {
     await createGroupController.showCreateOrEditGroupCohort(req, res)
   })
   getOrPost('/group/:groupId/edit-a-group/edit-group-cohort', async (req, res) => {
-    await editGroupCohortController.showEditGroupCohort(req, res)
+    await editGroupController.editGroupCohort(req, res)
   })
   getOrPost('/group/:groupId/edit-a-group/edit-group-sex', async (req, res) => {
-    await editGroupSexController.showEditGroupSex(req, res)
+    await editGroupController.editGroupSex(req, res)
   })
   getOrPost('/group/create-a-group/group-start-date', async (req, res) => {
     await createGroupController.showCreateGroupDate(req, res)
